Remove f-strings feature of python 3.6
F-strings are a new feature in python 3.6. If used they would force
an update to that python version. Everyone should always have the
latest version but its good to avoid imposing an update on users.

It would be bad to end up in a situation where users postpone updates
because they expect everything to break every time they update.