Seperate protocol handling to its own class
The Electrum protocol handling is now in its own file and class.
This is for better organization, and for later writing of tests
which test only the protocol handling code without needing to
create any sockets. The seperation will also be useful in future
when coding the feature to support multiple simultaneous client
connections.