Change logs to reduce spam and keep logfile small
The aim is to make the debug log file not grow as quickly. Now
there wont be messages printed every heartbeat. Also address_history
and the list of all imported addresses wont be written as they are
really huge and dont add much value.

The info log is also reduced with the aim to make it more useful
for users to watch. They will see incoming transactions.