Disconnect client on unknown scripthash
Previously the server would just print a warning to stdout and return
an empty wallet when the client requested an unknown scripthash. Users
dont always read stdout. It is better to actually disconnect the client
instead when this happens. This will make the user figure out what's
wrong and read the debug log or stdout.