Rewrite unconf tx handler to use getmempoolentry
Previously when handling an unconfirmed transaction the code
would use an ugly cascade of calls to `getrawtransaction`
and `gettxout` in an effort to learn whether an unconfirmed
had unconfirmed parents, and what its miner fee was.

This is replaced by using the `getmempoolentry` RPC call which
is much cleaner and shorter.

Update tests