Notify client the status of a tor broadcasted tx
See issue #220

Previously if Tor broadcasting was enabled the server would always reply
with a success message to the client as long as the transaction was valid,
regardless of what actually happened with the broadcasting.

One consequence was that if Tor was not running the broadcast would silently
fail, all the while displaying a success message to the user.

This commit uses a thread-safe queue to wait for a message from one of the
broadcasting threads. Once a thread uploads the transaction to the first peer
it will notify and wake up the server thread. Only then will the client see
a success message. For network failures or Tor not running a fail message
will be shown. There is also a timeout of 20 seconds after which a success
message will be shown, because Tor can be slow and theres no point making
the user stare at a "Please wait" dialog for so long.