Remove call to getaddressesbylabel
Previously the startup sequence would request all imported addresses and use them
to check whether addresses have already been imported or not. This would be slow
in situations where a large number of addresses were imported.

To improve startup time we now use the getaddressinfo RPC call instead.

I tried using Electrum Personal Server on a wallet which had one million addresses
imported as a test, and it was unusable for this reason.